build(support): Create support dir for development tools

+++ support/make_docs.sh
@@ -20,9 +20,25 @@ do
 done
 echo "done"
 
+echo -n "Updating Manifest..."
+if [[ -e 'MANIFEST' ]]; then
+  rm MANIFEST
+fi
+make manifest
+echo "done"
+
 # Update Changelog
 echo -n "Updating Changelog..."
-rm CHANGELOG.md
+if [[ -e 'CHANGELOG.md' ]]; then
+  rm CHANGELOG.md
+fi
 git cliff > CHANGELOG.md
+echo "done"
+
+echo -n "Updating Signatures..."
+if [[ -e 'SIGNATURE' ]]; then
+  rm SIGNATURE
+fi
+make signature
 echo "done"
